Cleaned up unit tests.
